delete roles with confirmation works

+++ app/javascript/components/Roles/Roles.js
@@ -41,21 +41,24 @@ const Roles = props => {
   const handleDestroy = (id, e) => {
     e.preventDefault()
 
-    const csrfToken = document.querySelector('[name=csrf-token]').content
-    axios.defaults.headers.common['X-CSRF-TOKEN'] = csrfToken
+    const confirmation = confirm("Do you really want to delete this item?");
+    if (confirmation) {
 
-    const url = `/api/v1/roles/${id}`
-    axios.delete(url)
-    .then( (data) => {
-      let roles = script.included.filter(item => item.type === "role")
-      console.log("Roles", roles)
-      const index = roles.findIndex(data => data.id == id)
-      console.log(index)
-      roles.splice(index, 1)
-      console.log("Roles", roles)
-      setScript({...script, roles})   // TODO: Update DOM after delete
-    })
-    .catch( data => console.log('Error', data) )
+      const csrfToken = document.querySelector('[name=csrf-token]').content
+      axios.defaults.headers.common['X-CSRF-TOKEN'] = csrfToken
+
+      const url = `/api/v1/roles/${id}`
+      axios.delete(url)
+      .then( (data) => {
+        let roles = script.included
+        const index = script.included.findIndex(data => data.id == id && data.type == "role")
+        roles.splice(index, 1)
+
+        setScript({...script, roles})
+
+      })
+      .catch( data => console.log('Error', data) )
+    }
   }
 
   let roles
